refactor: modularize maps/index.tsx
